requirements.txt
fastapi
uvicorn
pandas
requests
python-dotenv
openai
sqlalchemy
psycopg2-binary